úpravy podle feedbacku od Marka

+++ src/main/java/com/bezdekova/bookstore/repositories/AuthorRepository.java
@@ -3,6 +3,6 @@ package com.bezdekova.bookstore.repositories;
 import com.bezdekova.bookstore.db.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepository extends JpaRepository<Author, Long> {
+public interface AuthorRepository extends JpaRepository<Author, Integer> {
     
 }
